+++ src/module/vue/components/character-header.vue
@@ -1,21 +1,7 @@
 <template>
   <header class="sheet-header">
-    <img
-      :src="actor.img"
-      :title="actor.name"
-      class="profile-img"
-      data-edit="img"
-      height="50"
-      width="50"
-    />
-    <h1 class="charname">
-      <input
-        :placeholder="$t('IRONSWORN.Name')"
-        v-model="actor.name"
-        name="name"
-        type="text"
-      />
-    </h1>
+    <document-img :document="actor" />
+    <document-name :document="actor" />
     <div class="flexrow xp">
       <h4 style="flex-grow: 0; margin: 5px">{{ $t('IRONSWORN.XP') }}</h4>
       <div class="flexrow">
@@ -42,13 +28,6 @@ export default {
     return {
       xpArray: [1, 2, 3, 4, 5, 6, 7, 8, 9],
     }
-  },
-
-  watch: {
-    'actor.name': CONFIG.IRONSWORN._.debounce(function() {
-      const actor = game.actors?.get(this.actor._id)
-      actor.update({ name: this.actor.name })
-    }, 1000),
   },
 }
 </script>