+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Vue asset sheet ([#150](https://github.com/ben/foundry-ironsworn/pull/150))
 - Make the Vue sheets default for all types ([#152](https://github.com/ben/foundry-ironsworn/pull/152))
+- Vue cleanup ([#153](https://github.com/ben/foundry-ironsworn/pull/153))
 
 ## 1.7.1